requirements.txt
accelerate==1.12.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
babel==2.17.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
coloredlogs==15.0.1
confection==0.1.5
csvw==3.7.0
ctranslate2==4.6.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
datasets==4.4.1
dill==0.4.0
diskcache==5.6.3
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
exceptiongroup==1.3.1
fastapi==0.123.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
genson==1.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
interegular==0.3.3
iso3166==2.1.1
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kokoro==0.9.4
language-tags==1.2.0
lark==1.3.1
loguru==0.7.3
MarkupSafe==2.1.5
misaki==0.9.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
nest-asyncio==1.6.0
networkx==3.3
num2words==0.5.14
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.23.2
openwakeword==0.6.0
outlines==0.2.3
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
phonemizer-fork==3.3.2
pillow==11.3.0
ply==3.11
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyaml==25.7.0
pyarrow==22.0.0
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.29.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
segments==2.3.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speexdsp-ns==0.1.2
srsly==2.5.2
starlette==0.50.0
sympy==1.13.1
termcolor==2.5.0
tflite-runtime==2.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.3
triton==3.1.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
